requirements.txt
# requirements.txt - Fixed with complete dependencies

# Packet Capture
scapy==2.5.0
pyshark==0.6
nfstream==6.5.4
dpkt==1.9.8

# Streaming & Message Queue
kafka-python==2.0.2
confluent-kafka==2.3.0

# Caching & State
redis==5.0.1

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.1
xgboost==2.0.3
joblib==1.3.2

# Cloud SDKs
boto3==1.34.0
azure-mgmt-network==25.0.0
azure-identity==1.15.0
google-cloud-logging==3.8.0
google-cloud-compute==1.14.0

# Monitoring & Metrics
prometheus-client==0.19.0
influxdb-client==1.37.0

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
typing_extensions>=4.8.0
python-multipart==0.0.6

# Security
cryptography==41.0.7
pyjwt==2.8.0
passlib==1.7.4

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
aiohttp==3.9.1
requests==2.31.0
psutil==5.9.6
python-json-logger==2.0.7

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
